+++ example/lib/testd.dart
@@ -1,8 +1,10 @@
+import 'dart:convert';
+
 import 'package:example/example.dart';
 import 'package:example/gen/artifacts.gen.dart';
 
 void main() {
   Base v = Base(baseValue: 3);
 
-  print($Base.fromMap({"baseValue": 0}));
+  print(jsonEncode($Base.schema));
 }